load laterooms from file
git-svn-id: https://subversion.assembla.com/svn/importproducts/trunk@45 c09a0911-d79b-4797-ba2f-bc3321b6aa18

+++ ImportProducts/ImportLaterooms.cs
@@ -124,41 +124,40 @@ namespace ImportProducts
             string countryFilter = param.CountryFilter;
             string cityFilter = param.CityFilter;
 
-#if TEST
-            _URL = @"C:\Temp\Hotels_Standard.xml";
-#else
-            // unzip file to temp folder if needed
-            if (_URL.EndsWith(".zip"))
+            if (!File.Exists(_URL))
             {
-                string zipFileName = String.Format("{0}\\{1}", Properties.Settings.Default.TempPath,
-                                                   "Hotels_Standard.zip");
-                // inside this function show progressBar for step: LoadFile
-                if (File.Exists(zipFileName))
+                // unzip file to temp folder if needed
+                if (_URL.EndsWith(".zip"))
                 {
-                    File.Delete(zipFileName);
-                }
-                SaveFileFromURL(_URL, zipFileName, 60, bw, e);
-
-                // if user cancel during saving file or ERROR
-                if (e.Cancel || (e.Result != null) && e.Result.ToString().Substring(0, 6).Equals("ERROR:")) return;   
-                // Set step for backgroundWorker
-                Form1.activeStep = "Extract file..";
-                bw.ReportProgress(0);           // start new step of background process
-
-                using (ZipFile zip1 = ZipFile.Read(zipFileName))
-                {
-
-                    foreach (ZipEntry zipEntry in zip1)
+                    string zipFileName = String.Format("{0}\\{1}", Properties.Settings.Default.TempPath,
+                                                       "Hotels_Standard.zip");
+                    // inside this function show progressBar for step: LoadFile
+                    if (File.Exists(zipFileName))
                     {
-                        zipEntry.Extract(Properties.Settings.Default.TempPath,
-                                         ExtractExistingFileAction.OverwriteSilently);
-
-
+                        File.Delete(zipFileName);
                     }
+                    SaveFileFromURL(_URL, zipFileName, 60, bw, e);
+
+                    // if user cancel during saving file or ERROR
+                    if (e.Cancel || (e.Result != null) && e.Result.ToString().Substring(0, 6).Equals("ERROR:")) return;
+                    // Set step for backgroundWorker
+                    Form1.activeStep = "Extract file..";
+                    bw.ReportProgress(0); // start new step of background process
+
+                    using (ZipFile zip1 = ZipFile.Read(zipFileName))
+                    {
+
+                        foreach (ZipEntry zipEntry in zip1)
+                        {
+                            zipEntry.Extract(Properties.Settings.Default.TempPath,
+                                             ExtractExistingFileAction.OverwriteSilently);
+
+
+                        }
+                    }
+                    _URL = String.Format("{0}\\{1}", Properties.Settings.Default.TempPath, "Hotels_Standard.xml");
                 }
-                _URL = String.Format("{0}\\{1}", Properties.Settings.Default.TempPath, "Hotels_Standard.xml");
             }
-#endif
 
             // show progress & catch Cancel
             if (bw.CancellationPending)
